bugfix  - delete temp file  - encoding  - EnvelopeData/@contentDescription to  200 chars

+++ src/main/java/si/jrc/msh/plugin/meps/pdf/PDFContentData.java
@@ -16,38 +16,12 @@ import java.util.logging.Logger;
  */
 public class PDFContentData {
 
-    private byte[] mbaData = null;
+
     private int miDocumentCount = -1;
     private int miPageCount = -1;
     private String mstrTempFileName =  null;
 
-    public byte[] getData() {
-        byte[] ba = mbaData;
-        /*
-        //System.out.println("MailContentData: getData tempFile: " + mstrTempFileName);
-        if (ba == null && mstrTempFileName!= null){
-            FileInputStream fis = null;
-            try {
-                //System.out.println("MailContentData: read from tempFile: " + mstrTempFileName);
-                fis = new FileInputStream(mstrTempFileName);
-                ba = new byte[fis.available()];
-                fis.read(ba);
-            } catch (IOException ex) {
-                Logger.getLogger(PDFContentData.class.getName()).log(Level.SEVERE, null, ex);
-            }  finally {
-                try {
-                    fis.close();
-                } catch (IOException ex) {
-                    Logger.getLogger(PDFContentData.class.getName()).log(Level.SEVERE, null, ex);
-                }
-            }
-        }*/
-        return ba;
-    }
 
-    public void setData(byte[] mbaData) {
-        this.mbaData = mbaData;
-    }
 
     public int getDocumentCount() {
         return miDocumentCount;
@@ -65,32 +39,16 @@ public class PDFContentData {
         this.miPageCount = miPageCount;
     }
 
+    public void setTempFileName(String file) {
+        mstrTempFileName = file;
+    }
     public String getTempFileName() {
         return mstrTempFileName;
     }
 
-    public void setTempFileName(String mstrTempFileName) {
-        this.mstrTempFileName = mstrTempFileName;
-        if (this.mstrTempFileName!= null){
-            FileInputStream fis = null;
-            try {
-                //System.out.println("MailContentData: read from tempFile: " + mstrTempFileName);
-                fis = new FileInputStream(mstrTempFileName);
-                mbaData = new byte[fis.available()];
-                fis.read(mbaData);
-            } catch (IOException ex) {
-                Logger.getLogger(PDFContentData.class.getName()).log(Level.SEVERE, null, ex);
-            }  finally {
-                try {
-                    fis.close();
-                } catch (IOException ex) {
-                    Logger.getLogger(PDFContentData.class.getName()).log(Level.SEVERE, null, ex);
-                }
-            }
-        }
-    }
+   
     public void deleteTempFile(){
-        if(mstrTempFileName!=null){
+        if(mstrTempFileName!=null && !mstrTempFileName.isEmpty()){
             File f = new File(mstrTempFileName);
             if (f.exists()) f.delete();
         }
